:fix: Corrigindo o bug na ordenação das partições

+++ sorting/quick_sort.h
@@ -2,4 +2,4 @@
 
 #include "../big_file.h"
 
-void quickSort (ITEM_VENDA *itens, unsigned long int direita, unsigned long int esquerda);
+void quickSort (ITEM_VENDA *itens, int direita, int esquerda);
